+++ CHANGELOG.md
@@ -4,7 +4,15 @@
 converting data unit to physical unit.
 
 - Update error prediction to 8/sqrt(3) * rms/sqrt(N) based on the data bandwidth 100 Hz
+- Add method to ArrayMap class to read in grating configuration and converting array map to wavelength
+- Add function freq_to_wl() and wl_to_freq()
+- Add function transmission(), transmission_range(), transmission_smoothed(), transmission_smoothed_range(), and
+  transmission_window() to pipeline
+- Add set_xscale() and set_yscale() method to FigArray class
+- Add xscale parameter to FigArray.psd(), enabling to plot log-log power spectral diagram, and change the behaviour of
+  scale parameter
 - Fix a bug in zus2io.proc_along_axis()
+- Change db to 10*log10(power) instead of 20, because power instead of root power is used here
 
 1.6